Almost done. Error with adding Gift

+++ src/main/java/ru/tylyakov/urfuproject/service/GiftServiceImpl.java
@@ -1,6 +1,7 @@
 package ru.tylyakov.urfuproject.service;
 
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 import ru.tylyakov.urfuproject.entity.Gift;
 import ru.tylyakov.urfuproject.repository.GiftRepository;
@@ -19,12 +20,32 @@ public class GiftServiceImpl implements GiftService {
     }
 
     @Override
-    public void save(Gift gift) {
+    public void save(Gift gift, Authentication authentication) {
+
+        // Проверка прав доступа
+        boolean isReadOnly = authentication.getAuthorities()
+                .stream()
+                .anyMatch(a -> a.getAuthority().equals("ROLE_READ_ONLY"));
+
+        if (isReadOnly) {
+            throw new RuntimeException("У вас нет прав для создания подарков. Обратитесь к администратору для повышения роли.");
+        }
+
         giftRepository.save(gift);
     }
 
     @Override
-    public void delete(Long id) {
+    public void delete(Long id, Authentication authentication) {
+
+        // Проверка прав доступа
+        boolean isReadOnly = authentication.getAuthorities()
+                .stream()
+                .anyMatch(a -> a.getAuthority().equals("ROLE_READ_ONLY"));
+
+        if (isReadOnly) {
+            throw new RuntimeException("У вас нет прав для удаления подарков. Обратитесь к администратору для повышения роли.");
+        }
+
         giftRepository.deleteById(id);
     }
 }
